feat: restructure and embed frontend

+++ pkg/upload/s3.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"context"
 	"fmt"
+	"io"
 	"net/http"
 	"net/url"
 	"path"
@@ -16,17 +17,20 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/aws/aws-sdk-go-v2/service/s3/types"
 	transport "github.com/aws/smithy-go/endpoints"
+	"github.com/gabriel-vasile/mimetype"
 	"github.com/gopad/gopad-api/pkg/config"
 )
 
 // S3Upload implements the Upload interface.
 type S3Upload struct {
-	endpoint string
-	path     string
-	access   string
-	secret   string
-	bucket   string
-	region   string
+	endpoint  string
+	path      string
+	access    string
+	secret    string
+	bucket    string
+	region    string
+	pathstyle bool
+	proxy     bool
 
 	client  *s3.Client
 	presign *s3.PresignClient
@@ -40,6 +44,8 @@ func (u *S3Upload) Info() map[string]interface{} {
 	result["path"] = u.path
 	result["bucket"] = u.bucket
 	result["region"] = u.region
+	result["pathstyle"] = u.pathstyle
+	result["proxy"] = u.proxy
 
 	return result
 }
@@ -86,7 +92,11 @@ func (u *S3Upload) Prepare() (Upload, error) {
 		return nil, fmt.Errorf("failed to load configuration: %w", err)
 	}
 
-	serviceOpts := []func(*s3.Options){}
+	serviceOpts := []func(*s3.Options){
+		func(o *s3.Options) {
+			o.UsePathStyle = u.pathstyle
+		},
+	}
 
 	if u.endpoint != "" {
 		endpoint, err := url.Parse(u.endpoint)
@@ -99,10 +109,25 @@ func (u *S3Upload) Prepare() (Upload, error) {
 			serviceOpts,
 			s3.WithEndpointResolverV2(
 				&CustomEndpointResolver{
-					Endpoint: endpoint,
+					Endpoint:  endpoint,
+					Bucket:    u.bucket,
+					PathStyle: u.pathstyle,
 				},
 			),
 		)
+
+		if endpoint.Scheme == "http" {
+			serviceOpts = append(
+				serviceOpts,
+				func(o *s3.Options) {
+					o.HTTPClient = &http.Client{
+						Transport: &http.Transport{
+							TLSClientConfig: nil,
+						},
+					}
+				},
+			)
+		}
 	}
 
 	u.client = s3.NewFromConfig(
@@ -123,13 +148,25 @@ func (u *S3Upload) Close() error {
 }
 
 // Upload stores an attachment within the defined S3 bucket.
-func (u *S3Upload) Upload(ctx context.Context, key, ctype string, content []byte) error {
+func (u *S3Upload) Upload(ctx context.Context, key string, content *bytes.Buffer) error {
+	reader := bytes.NewReader(
+		content.Bytes(),
+	)
+
+	mtype, err := mimetype.DetectReader(
+		reader,
+	)
+
+	if err != nil {
+		return err
+	}
+
 	params := &s3.PutObjectInput{
 		ACL:         types.ObjectCannedACLPublicRead,
 		Bucket:      aws.String(u.bucket),
 		Key:         aws.String(path.Join(u.path, key)),
-		ContentType: aws.String(ctype),
-		Body:        bytes.NewReader(content),
+		ContentType: aws.String(mtype.String()),
+		Body:        reader,
 	}
 
 	if _, err := u.client.PutObject(
@@ -143,17 +180,71 @@ func (u *S3Upload) Upload(ctx context.Context, key, ctype string, content []byte
 }
 
 // Delete removes an attachment from the defined S3 bucket.
-func (u *S3Upload) Delete(ctx context.Context, key string) error {
-	params := &s3.DeleteObjectInput{
-		Bucket: aws.String(u.bucket),
-		Key:    aws.String(path.Join(u.path, key)),
-	}
+func (u *S3Upload) Delete(ctx context.Context, key string, recursive bool) error {
+	if recursive {
+		var (
+			continuation *string
+		)
 
-	if _, err := u.client.DeleteObject(
-		ctx,
-		params,
-	); err != nil {
-		return err
+		for {
+			objects, err := u.client.ListObjectsV2(
+				ctx, &s3.ListObjectsV2Input{
+					Bucket:            aws.String(u.bucket),
+					Prefix:            aws.String(path.Join(u.path, key)),
+					ContinuationToken: continuation,
+				},
+			)
+
+			if err != nil {
+				return fmt.Errorf("failed to list objects: %w", err)
+			}
+
+			if len(objects.Contents) == 0 {
+				break
+			}
+
+			deletable := make([]types.ObjectIdentifier, 0)
+
+			for _, object := range objects.Contents {
+				deletable = append(
+					deletable,
+					types.ObjectIdentifier{
+						Key: object.Key,
+					},
+				)
+			}
+
+			if _, err = u.client.DeleteObjects(
+				ctx,
+				&s3.DeleteObjectsInput{
+					Bucket: aws.String(u.bucket),
+					Delete: &types.Delete{
+						Objects: deletable,
+						Quiet:   aws.Bool(true),
+					},
+				},
+			); err != nil {
+				return fmt.Errorf("failed to delete objects: %w", err)
+			}
+
+			if !aws.ToBool(objects.IsTruncated) {
+				break
+			}
+
+			continuation = objects.NextContinuationToken
+		}
+	} else {
+		params := &s3.DeleteObjectInput{
+			Bucket: aws.String(u.bucket),
+			Key:    aws.String(path.Join(u.path, key)),
+		}
+
+		if _, err := u.client.DeleteObject(
+			ctx,
+			params,
+		); err != nil {
+			return err
+		}
 	}
 
 	return nil
@@ -162,44 +253,120 @@ func (u *S3Upload) Delete(ctx context.Context, key string) error {
 // Handler implements an HTTP handler for asset uploads.
 func (u *S3Upload) Handler(root string) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		req, err := u.presign.PresignGetObject(
-			r.Context(),
-			&s3.GetObjectInput{
-				Bucket: aws.String(u.bucket),
-				Key:    aws.String(path.Join(u.path, strings.TrimPrefix(r.URL.Path, root))),
-			},
-			func(opts *s3.PresignOptions) {
-				opts.Expires = time.Duration(5 * time.Minute)
-			},
+		if u.proxy {
+			u.proxyHandler(root, w, r)
+		} else {
+			u.presignHandler(root, w, r)
+		}
+	})
+}
+
+func (u *S3Upload) proxyHandler(root string, w http.ResponseWriter, r *http.Request) {
+	obj := strings.TrimPrefix(
+		path.Join(
+			u.path,
+			strings.TrimPrefix(
+				r.URL.Path,
+				root,
+			),
+		),
+		"/",
+	)
+
+	req, err := u.client.GetObject(
+		r.Context(),
+		&s3.GetObjectInput{
+			Bucket: aws.String(u.bucket),
+			Key:    aws.String(obj),
+		},
+	)
+
+	if err != nil {
+		http.Error(
+			w,
+			http.StatusText(http.StatusNotFound),
+			http.StatusNotFound,
 		)
 
-		if err != nil {
-			http.Error(
-				w,
-				http.StatusText(http.StatusNotFound),
-				http.StatusNotFound,
-			)
-		}
+		return
+	}
 
-		http.Redirect(w, r, req.URL, http.StatusTemporaryRedirect)
-	})
+	defer req.Body.Close()
+
+	modified := time.Now()
+
+	if req.LastModified != nil {
+		modified = *req.LastModified
+	}
+
+	buffer, err := io.ReadAll(req.Body)
+
+	if err != nil {
+		http.Error(
+			w,
+			"Failed to read buffer",
+			http.StatusInternalServerError,
+		)
+
+		return
+	}
+
+	http.ServeContent(
+		w,
+		r,
+		obj,
+		modified,
+		bytes.NewReader(buffer),
+	)
+}
+
+func (u *S3Upload) presignHandler(root string, w http.ResponseWriter, r *http.Request) {
+	obj := strings.TrimPrefix(
+		path.Join(
+			u.path,
+			strings.TrimPrefix(
+				r.URL.Path,
+				root,
+			),
+		),
+		"/",
+	)
+
+	req, err := u.presign.PresignGetObject(
+		r.Context(),
+		&s3.GetObjectInput{
+			Bucket: aws.String(u.bucket),
+			Key:    aws.String(obj),
+		},
+		func(opts *s3.PresignOptions) {
+			opts.Expires = time.Duration(5 * time.Minute)
+		},
+	)
+
+	if err != nil {
+		http.Error(
+			w,
+			http.StatusText(http.StatusNotFound),
+			http.StatusNotFound,
+		)
+
+		return
+	}
+
+	http.Redirect(w, r, req.URL, http.StatusTemporaryRedirect)
 }
 
 // NewS3Upload initializes a new S3 handler.
 func NewS3Upload(cfg config.Upload) (Upload, error) {
-	path := "/"
-
-	if cfg.Path != "" {
-		path = cfg.Path
-	}
-
 	f := &S3Upload{
-		endpoint: cfg.Endpoint,
-		path:     path,
-		access:   cfg.Access,
-		secret:   cfg.Secret,
-		bucket:   cfg.Bucket,
-		region:   cfg.Region,
+		endpoint:  cfg.Endpoint,
+		pathstyle: cfg.Pathstyle,
+		path:      cfg.Path,
+		access:    cfg.Access,
+		secret:    cfg.Secret,
+		bucket:    cfg.Bucket,
+		region:    cfg.Region,
+		proxy:     cfg.Proxy,
 	}
 
 	return f.Prepare()
@@ -218,12 +385,22 @@ func MustS3Upload(cfg config.Upload) Upload {
 
 // CustomEndpointResolver is used for S3 compatible storage endpoints.
 type CustomEndpointResolver struct {
-	Endpoint *url.URL
+	Endpoint  *url.URL
+	Bucket    string
+	PathStyle bool
 }
 
 // ResolveEndpoint resolves endpoints for a specific service and region
 func (r *CustomEndpointResolver) ResolveEndpoint(_ context.Context, _ s3.EndpointParameters) (transport.Endpoint, error) {
+	endpoint := r.Endpoint
+
+	if r.PathStyle {
+		endpoint = endpoint.JoinPath(
+			r.Bucket,
+		)
+	}
+
 	return transport.Endpoint{
-		URI: *r.Endpoint,
+		URI: *endpoint,
 	}, nil
 }
